Move it back to the lib folder
I mean, ts isn't es, so it makes more sense. Only put it in es bc I
misunderstood something about typedoc
