+++ src/lib/pixelmanipulator.ts
